NexusEngine/requirements.txt
# Python Dependencies for NexusEngine Omega

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
typer==0.9.0
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1

# Caching & Message Queue
redis==5.0.1
httpx==0.25.2

# Performance & Optimization
numpy==1.24.3
scipy==1.11.4
Cython==3.0.5

# Data Processing
pandas==2.1.3
polars==0.19.12

# Monitoring & Observability
prometheus-client==0.19.0
python-json-logger==2.0.7
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-prometheus==0.42b0
opentelemetry-instrumentation-fastapi==0.42b0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
hypothesis==6.92.1

# Code Quality
black==23.12.0
isort==5.13.2
flake8==6.1.0
pylint==3.0.3
mypy==1.7.1

# Development
ipython==8.18.1
jupyter==1.0.0
ipdb==0.13.13

# API Documentation
mkdocs==1.5.3
mkdocs-material==9.4.14

# Cryptography
cryptography==41.0.7

# Compression
zstandard==0.22.0

# Rate Limiting
slowapi==0.1.9

# Task Queue
celery==5.3.4
